WL-3705 Just tool in the tool for the popup value.
This means that when someone changes the popup value in a systemwide tool it won’t change straight away, but only after someone uses the LTI tool in a site. It’s not very nice but it’s the smallest fix.

The larger fix would have been to either to completely re-architect how LTI registers tools (not practical) or to have a DB job to change all uses of the tool when a change was made. This would have been possible but it’s a different style to how the rest of the LTI tool works.